Split make_model index and show tests
