+++ test/controllers/make_models_controller/index_test.rb
@@ -18,10 +18,10 @@ require 'shared/make_models_data'
 #     </ul>
 #   </dd>
 # </dl>
-class MakeModelsControllerTest < ActionDispatch::IntegrationTest
+class MakeModelsController::IndexTest < ActionDispatch::IntegrationTest
   include MakeModelsData
 
-  def test_index
+  test 'index view' do
     mmods = setup_make_models_data
     get make_models_url
     assert_response :success
@@ -48,28 +48,5 @@ class MakeModelsControllerTest < ActionDispatch::IntegrationTest
     assert_select('dl.make-models dt.make', mm.make)
     assert_select('dl.make-models dd.make ul.models li.model',
       "{no model specified}")
-  end
-
-  # <h1>S1T</h1>
-  # <dl class="make-model">
-  #   <dt>Make</dt><dd>Pitts</dd>
-  #   <dt>Model</dt><dd>S1T</dd>
-  #   <dt>Empty weight</dt><dd>780 lbs</dd>
-  #   ...
-  # <dl>
-  def test_show
-    mm = create(:make_model)
-    get make_model_url(mm)
-    assert_response :success
-    assert_select('h1', mm.model)
-    assert_select('dl.make-model', 1)
-    %w[Make Model Empty\ weight Maximum\ weight
-        Horsepower Seats Wings].each do |attrib|
-      assert_select('dt', attrib)
-    end
-    %w[make model empty_weight_lbs max_weight_lbs
-        horsepower seats wings].each do |attrib|
-      assert_select('dd', /#{mm.send(attrib)}/)
-    end
   end
 end
